Make code more DRY by extend all tree classes from FruitTree class
