+++ mango_tree.js
@@ -1,46 +1,8 @@
-class MangoTree {
+const FruitTree = require('./fruit_tree');
+
+class MangoTree extends FruitTree{
   constructor(name, age, height, matureAge, healthyStatus) {
-    this._name = name;
-    this._age = age;
-    this._height = height;
-    this._matureAge = matureAge;
-    this._healthyStatus = healthyStatus;
-    this._maxAge = 20;
-  }
-
-  get name() {
-    return this._name;
-  }
-
-  get age() {
-    return this._age;
-  }
-
-  get height() {
-    return this._height;
-  }
-
-  get matureAge() {
-    return this._matureAge;
-  }
-
-  get healthyStatus() {
-    return this._healthyStatus;
-  }
-
-  get maxAge() {
-    return this._maxAge;
-  }
-
-  grow() {
-    this._age += 1;
-    if (this._age < 8) {
-      const growth = Math.round(Math.random() * 10) / 10;
-      this._height += growth;
-    }
-    if (this._age === this._maxAge) {
-      this._healthyStatus = false;
-    }
+    super(name, age, height, matureAge, healthyStatus, 20);
   }
 }
